moved beginning animation to onboarding view controller
